Add shuffle_after_epoch to WebDataset, TFRecord, and MXNet readers (#6288)
  This PR adds `shuffle_after_epoch` and `shuffle_after_epoch_seed` support to the **WebDataset**,
  **TFRecord**, and **MXNet** readers - completing the file/shard-level shuffle feature that was
  previously added for file-based readers (PR #6236).

  When `shuffle_after_epoch=True`, the reader reshuffles the order of whole files/tar archives
  between epochs while preserving sequential I/O within each file. This is complementary to
  `random_shuffle`, which shuffles samples within the prefetch buffer -  both options can be
  combined. `shuffle_after_epoch_seed` controls reproducibility of the file-order shuffle.

  In addition, two pre-existing bugs were fixed that were only observable once file-order shuffling
  was in place:

  - **`IndexedFileLoader` (TFRecord)**: `should_seek_` was not set to `true` after reopening a file,
    causing reads at wrong positions when `next_seek_pos_` happened to equal the first record offset
    of the new file.
  - **`RecordIOLoader` (MXNet)**: `ReadSample` did not reopen the file when `file_index` changed,
    so multi-file RecordIO setups would read from the wrong file.

+++ dali/operators/reader/fits_reader_op.cc
@@ -71,7 +71,7 @@ This argument is mutually exclusive with `files`.)",
 
 `stick_to_shard` and `random_shuffle` cannot be used when this argument is set to True.)",
                     false)
-    .AddOptionalArg<int32_t>("shuffle_after_epoch_seed",
+    .AddOptionalArg<int64_t>("shuffle_after_epoch_seed",
                     R"(Random seed for the dataset shuffling performed after each epoch.
 
 If not provided, a fixed default seed is used, which results in the same shuffling
