+++ dali/operators/reader/loader/nemo_asr_loader.h
@@ -114,7 +114,7 @@ class DLL_PUBLIC NemoAsrLoader : public Loader<CPUBackend, AsrSample, true> {
         manifest_filepaths_(spec.GetRepeatedArgument<std::string>("manifest_filepaths")),
         shuffle_after_epoch_(spec.GetArgument<bool>("shuffle_after_epoch")),
         shuffle_after_epoch_seed_([&spec]() {
-          int32_t seed = kDaliDataloaderSeed;
+          int64_t seed = kDaliDataloaderSeed;
           spec.TryGetArgument(seed, "shuffle_after_epoch_seed");
           return seed;
         }()),
@@ -180,7 +180,7 @@ class DLL_PUBLIC NemoAsrLoader : public Loader<CPUBackend, AsrSample, true> {
   std::vector<size_t> shuffled_indices_;
 
   bool shuffle_after_epoch_;
-  int32_t shuffle_after_epoch_seed_;
+  int64_t shuffle_after_epoch_seed_;
   Index current_index_ = 0;
   int current_epoch_ = 0;
